move reachability into its own file
